+++ src/wildboar/explain/counterfactual/__init__.py
@@ -9,6 +9,7 @@ methods for generating counterfactual explanations.
 from ._helper import _proximity as proximity
 from ._helper import counterfactuals
 from ._nn import KNeighborsCounterfactual
+from ._nun import NativeGuideCounterfactual
 from ._proto import PrototypeCounterfactual
 from ._sf import ShapeletForestCounterfactual
 
@@ -18,4 +19,5 @@ __all__ = [
     "ShapeletForestCounterfactual",
     "KNeighborsCounterfactual",
     "PrototypeCounterfactual",
+    "NativeGuideCounterfactual",
 ]